+++ ethereum-wallet/Classes/Business layer/Models/RealmToken.swift
@@ -23,7 +23,7 @@ class RealmToken: Object {
   @objc dynamic var name: String = ""
   @objc dynamic var iso: String = ""
   @objc dynamic var address: String = ""
-  @objc dynamic var decimals: Int64 = 0
+  @objc dynamic var decimals: Int = 0
   @objc dynamic var lastUpdateTime: Date = .distantPast
   var rates = List<RealmRate>()
